adjust editor_cpp

+++ EDITOR_CPP/toolset/unicode.cpp
@@ -1,6 +1,6 @@
-#include "utf8.hh"
+#include "unicode.hh"
 
-size_t utf8get(const char *ptr, const char *end) {
+int utf8get(const char *ptr, const char *end) {
     int size = 0;
     
     //check first byte
