+++ EDITOR_CPP/toolset/unicode.hh
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "std.hh"
+#include "stdlibs.hh"
 
 static const char utf8bom[] = {'\xEF', '\xBB', '\xBF'};
 
 //return size of character; if not utf8 character, return 0.
-size_t utf8get(const char *ptr, const char *end);
+int utf8get(const char *ptr, const char *end);
